Simplified async responses processing
