+++ Api/Services/SupplierResponses/WebhookResponseService.cs
@@ -1,23 +1,18 @@
 ﻿using System.IO;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Services.Accommodations.Bookings;
-using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.ResponseProcessing;
 using HappyTravel.Edo.Api.Services.Connectors;
 using HappyTravel.Edo.Common.Enums;
 
-namespace HappyTravel.Edo.Api.Services.ProviderResponses
+namespace HappyTravel.Edo.Api.Services.SupplierResponses
 {
     public class WebhookResponseService
     {
-        public WebhookResponseService(
-             ISupplierConnectorManager supplierConnectorManager,
-             IBookingRecordsManager bookingRecordsManager,
+        public WebhookResponseService(ISupplierConnectorManager supplierConnectorManager,
              IBookingResponseProcessor responseProcessor)
         {
             _supplierConnectorManager = supplierConnectorManager;
-            _bookingRecordsManager = bookingRecordsManager;
             _responseProcessor = responseProcessor;
         }
         
@@ -28,17 +23,11 @@ namespace HappyTravel.Edo.Api.Services.ProviderResponses
             if (isGettingBookingDetailsFailure)
                 return Result.Failure(gettingBookingDetailsError.Detail);
             
-            var (_, isGetBookingFailure, booking, getBookingError) = await _bookingRecordsManager.Get(bookingDetails.ReferenceCode);
-            
-            if (isGetBookingFailure)
-                return Result.Failure(getBookingError);
-            
             await _responseProcessor.ProcessResponse(bookingDetails);
             return Result.Success();
         }
 
         private readonly ISupplierConnectorManager _supplierConnectorManager;
-        private readonly IBookingRecordsManager _bookingRecordsManager;
         private readonly IBookingResponseProcessor _responseProcessor;
     }
 }
